Initial work on bigbananajour
Bigbananajour is bananajour with an environment variable set that lets
it know that it is in 'mirror mode'.

In mirror mode, it will start a bananajour instance (web server and git
daemons) but listening on different ports to avoid conflicting with a
regular bananajour instance running on the same machine.

Bigbananajour instances also advertise themselves with 'ismirror'
property equal to true.  This is so that if more than one bigbananajour
instance is started on the same network they won't try to mirror each
other or any crazy shit like that. Regular bananajours have ismirror =
false.